Aggregated map didn't work. Debugging.

+++ analyses/aggregate_results_to_10_km.py
@@ -174,7 +174,7 @@ def aggregate(tile, thresh):
     # from the 2D array created by rasterio above
     # https://gis.stackexchange.com/questions/279953/numpy-array-to-gtiff-using-rasterio-without-source-raster
     aggregated = rasterio.open("{0}_{1}_10km.tif".format(tile_id, tile_type), 'w',
-                                driver='GTiff', compress='lzw', nodata='0', dtype='float32', count=1,
+                                driver='GTiff', compress='lzw', nodata='0', dtype='float64', count=1,
                                 height=100, width=100,
                                 # pixelSizeY='0.1', pixelSizeX='0.1', height=100, width=100,
                                 crs='EPSG:4326', transform=from_origin(xmin,ymax,0.1,0.1))
